Implement PE DOS stub/header writing
In order to create a UEFI bootloader, we need to be able to construct a
PE32+ file. PE32+ files (thanks to Microsoft) start with a DOS header
and program. The most commonly used program in this context is a stub
that simply outputs 'This program cannot be run in DOS mode' and exits.
Hence, that is what we do here, too.

It probably wasn't necessary to do all of this: we could have simply
crafted a static 'DOS preamble' sort of thing and embedded that, but I
thought it was a nice exercise to prepare for writing the real meat here
(the PE file).