Refactor Executable interface into WriterTo
A Reader is definitely the wrong interface for this scenario: generally
we want to be writing the sections somewhere, and this makes the code
much, much simpler. For example, rather than having to keep track of
state and open handles etc. with the virtual sections, we can simply
write all of the real ELF sections in one function body.

Additionally, fix a bug where the ELF sections weren't being placed at
the locations we expect. This was a padding issue, as the real ELF
section data can be shorter than the gaps between real ELF sections in a
virtual section. The virtualSectionReader wasn't accounting for this;
this fix is much, much easier in a WriterTo world, which was the driver
behind this refactor.

I think that was the last bug stopping this thing from working -- and
the UEFI application now runs!!