Fix mod dependency resolution order and uniqueness
GRUB requires that modules appear in the PE file in dependency order,
i.e. such that all dependencies of a given module appear **before** the
module itself.

The dependency resolver was previously doing the opposite of this, which
caused GRUB to abort with a red herring of an error ('variable "prefix"
isn't set')

Additionally, we were including duplicate modules, since module
dependencies are not mutually exclusive sets (i.e. two modules can share
a dependency).