Improve handling of header size and uint32/64s
Previously, we were hardcoding the header size in efipe. This doesn't
make sense, as we specify it when creating the grub image; hence, the
two sizes could be out of sync, which would lead to broken images.

Additionally, slightly improve the handling of 32/64 bit ints. This
could still do with some further work. PE32+ images are limited to 2
gigabytes in size, but the ImageBase field is 64 bits. This latter point
is irrelevant to us, because we set this field to 0, as we don't need to
locate the image in 64-bit memory. This image runs so early in the boot
process that the odds of it being loaded in 64-bit memory are relatively
slim anyway!