Implement rudimentary ISO creation
This involved a refactor of cmd/ into something closer to what the
command-line application is likely to look like in its final form: a
single command with subcommands and a config.

It may seem strange that we're implementing ISOs before we even have
TFTP -- given this is intended to be PXE boot software -- but this
should hopefully help me iterate faster on the features that are
important for TFTP/PXE (such as distro acquisition).